+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Main (unreleased)
 
 - (_Experimental_) Additions to experimental `database_observability.mysql` component:
   - `query_sample` collector now supports auto-enabling the necessary `setup_consumers` settings (@cristiangreco)
+  - add `query_tables` collector for postgres (@matthewnolf)
 
 ### Enhancements